+++ marketplace/src/main/java/store/carjava/marketplace/app/marketplace_car/repository/MarketplaceCarCustomRepository.java
@@ -15,4 +15,8 @@ public interface MarketplaceCarCustomRepository {
                                     List<Long> optionIds, String testDriveCenterName, String status, Integer minEngineCapacity,
                                     Integer maxEngineCapacity, String name, String sortOrder , Pageable pageable
                                     );
+
+    List<MarketplaceCar> findMarketplaceCarProperList(long budgetLow, long budgetHigh, String vehicle);
+    MarketplaceCar findUpgradeModelCarOverPrice(long budget, String vehicle, MarketplaceCar marketplaceCar);
+    MarketplaceCar findCarMoreOptionOverPrice(long budget, String vehicle, MarketplaceCar marketplaceCar);
 }
